% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oDAGanno.r
\name{oDAGanno}
\alias{oDAGanno}
\title{Function to generate a DAG subgraph induced by the input annotation data}
\usage{
oDAGanno(
ig,
anno,
path.mode = c("all_paths", "shortest_paths", "all_shortest_paths",
"shortest_paths_epath"),
true.path.rule = TRUE,
verbose = TRUE
)
}
\arguments{
\item{ig}{an object of class "igraph" to represent DAG}

\item{anno}{the vertices/nodes for which annotation data are provided.
It can be a sparse Matrix of class "dgCMatrix" (with variants/genes as
rows and terms as columns), or a list of nodes/terms each containing
annotation data, or an object of class 'SET'}

\item{path.mode}{the mode of paths induced by vertices/nodes with input
annotation data. It can be "all_paths" for all possible paths to the
root, "shortest_paths" for only one path to the root (for each node in
query), "all_shortest_paths" for all shortest paths to the root (i.e.
for each node, find all shortest paths with the equal lengths),
"shortest_paths_epath" (indeed only edges in the shortest path are
kept)}

\item{true.path.rule}{logical to indicate whether the true-path rule
should be applied to propagate annotations. By default, it sets to
true}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
NULL or an induced subgraph, an object of class "igraph". In addition
to the original attributes to nodes and edges, the return subgraph is
also appended by two node attributes: 1) "anno" containing a list of
variants/genes either as original annotations (and inherited
annotations; 2) "IC" standing for information content defined as
negative 10-based log-transformed frequency of variants/genes annotated
to that term.
}
\description{
\code{oDAGanno} is supposed to produce a subgraph induced by the input
annotation data, given a direct acyclic graph (DAG; an ontology). The
input is a graph of "igraph", a list of the vertices containing
annotation data, and the mode defining the paths to the root of DAG.
The induced subgraph contains vertices (with annotation data) and their
ancestors along with the defined paths to the root of DAG. The
annotations at these vertices (including their ancestors) can also be
updated according to the true-path rule: those annotated to a term
should also be annotated by its all ancestor terms.
}
\note{
For the mode "shortest_paths", the induced subgraph is the most
concise, and thus informative for visualisation when there are many
nodes in query, while the mode "all_paths" results in the complete
subgraph.
}
\examples{
\dontrun{
###########################################################
g <- oRDS('ig.GOMF', placeholder=placeholder)
anno <- oRDS('org.Hs.egGOMF', placeholder=placeholder)
dag <- oDAGanno(g, anno)
}
}
\seealso{
\code{\link{oDAGinduce}}
}
