% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oDAGclade.r
\name{oDAGclade}
\alias{oDAGclade}
\title{Function to generate clades from DAG}
\usage{
oDAGclade(ig, nodes_query = NULL, verbose = TRUE)
}
\arguments{
\item{ig}{an object of class "igraph" to represent DAG}

\item{nodes_query}{nodes in query}

\item{verbose}{logical to indicate whether the messages will be
displayed in the screen. By default, it sets to true for display}
}
\value{
NULL or an object of class "igraph" or a list of "igraph" objects
}
\description{
\code{oDAGclade} is supposed to generate clades from a direct acyclic
graph (DAG; an ontology), that is, subgraphs induced by each of input
nodes.
}
\note{
none
}
\examples{
\dontrun{
###########################################################
ig <- oRDS('ig.GOMF', placeholder=placeholder)
nodes_query <- ig \%>\% oIG2TB('nodes') \%>\% filter(distance==1) \%>\%
pull(name)
clade <- oDAGclade(ig, nodes_query)
}
}
\seealso{
\code{\link{oDAGclade}}
}
