% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oPolarBar.r
\name{oPolarBar}
\alias{oPolarBar}
\title{Function to visualise a data frame using a polar barplot}
\usage{
oPolarBar(
df,
colormap = "spectral",
size.name = 10,
size.value = 3,
parallel = FALSE
)
}
\arguments{
\item{df}{a data frame with two columns ('name' and 'value')}

\item{colormap}{either NULL or color names ('spectral' by default) for
bars according to the name column}

\item{size.name}{an integer specifying the text size for the name
column. By default, it sets to 10}

\item{size.value}{an integer specifying the text size for the value
column. By default, it sets to 3}

\item{parallel}{logical to indicate whether the label is parallel to
polar coordinate. By default, it sets FALSE}
}
\value{
an object of class "ggplot"
}
\description{
\code{oPolarBar} is supposed to visualise a data frame using a polar
dotplot. It returns an object of class "ggplot".
}
\note{
none
}
\examples{
\dontrun{
# a) provide the seed nodes/genes with the weight info
## get genes within 500kb away from AS GWAS lead SNPs
seeds.genes <- ImmunoBase$AS$genes_variants
## seeds weighted according to distance away from lead SNPs
data <- 1- seeds.genes/500000

# b) prepare a data frame
df <- data.frame(name=names(data), value=data, stringsAsFactors=FALSE)

# c) do correlation
gp <- oPolarBar(df[1:20,], parallel=TRUE)
gp
}
}
\seealso{
\code{\link{oPolarBar}}
}
